Script that does everything
This script launches everything that needs to be launched on our machine, then launches
a server (it does rmiregistry too) on the other machine.

Since the ssh connection stays open, the output of both servers will be in the window that
launched ./prepare_local, which now that I think about it, should be called prepare_all.