Plans for secure and unsecured dispatching.
With secured we divide the work into pieces, one per server, the compute
callable breaks it up into bite-sized chunks, send the chunks and merges
the results.  Then the dispatcher merges the results from each server.

With unsecured, the dispatcher breaks the whole list into bite-sized
chunks.  For each chunk it dispatches to all the the servers
computeCallabbles and looks what each server answered to find out which
answer can be trusted.